Add files via upload
Write a program that uses using Backtracking Search to solves the problem of placing k knights on an n×n chessboard such that no two knights are attacking each other. The program should either return a valid placement, or FALSE if no such placement is possible (because k is too large in relation to n).